refactor(test): split oversized sync/graph test harnesses
Extract shared graph fixtures and sync run helpers instead of bumping
file-length exemptions. graph.test.ts uses graphIssue() + helpers;
sync integration tests move to focused sync-run-*.test.ts files with
shared mocks in sync-test-mocks.ts.
