+++ worker/src/api/graph-test-helpers.ts
@@ -1,4 +1,60 @@
-import type { FakeResult } from "../test-utils";
+import { Hono } from "hono";
+import type { AuthEnv } from "../auth/types";
+import {
+  type FakeResult,
+  STUB_SESSION,
+  captureDb,
+  dispatchByTable,
+  makeEnv,
+  makeFakeDb,
+  makeFakeStmt,
+} from "../test-utils";
+import type { Env } from "../types";
+import { graphRoute } from "./graph";
+
+export const GRAPH_REPO = "Roxabi/roxabi-live";
+
+export type GraphIssueRow = {
+  key: string;
+  repo: string;
+  number: number;
+  title: string | null;
+  state: string;
+  url: string | null;
+  milestone: null;
+  lane: string | null;
+  priority: string | null;
+  size: string | null;
+  status: string | null;
+  is_stub: number;
+  has_active_branch: number;
+  assignees?: string;
+  updated_at?: string | null;
+};
+
+/** Minimal issues-row fixture for graph route tests. */
+export function graphIssue(
+  number: number,
+  overrides: Partial<GraphIssueRow> = {},
+  repo = GRAPH_REPO,
+): GraphIssueRow {
+  return {
+    key: `${repo}#${number}`,
+    repo,
+    number,
+    title: null,
+    state: "open",
+    url: null,
+    milestone: null,
+    lane: null,
+    priority: null,
+    size: null,
+    status: null,
+    is_stub: 0,
+    has_active_branch: 0,
+    ...overrides,
+  };
+}
 
 export function aggregateRepoActivity(
   issues: Array<{ repo: string; updated_at?: string | null }>,
@@ -17,4 +73,94 @@ export function aggregateRepoActivity(
     }
   }
   return [...byRepo.entries()].map(([repo, stats]) => ({ repo, ...stats }));
+}
+
+export function makeGraphTestApp() {
+  const a = new Hono<AuthEnv>();
+  a.use("*", async (c, next) => {
+    c.set("session", STUB_SESSION);
+    await next();
+  });
+  a.get("/api/graph", graphRoute);
+  return a;
+}
+
+function graphDbRows(
+  sql: string,
+  labels: unknown[],
+  prState: unknown[],
+  issues: unknown[],
+  edges: unknown[],
+  repos: unknown[],
+  visibleRepos: string[],
+  zkOptIn: boolean,
+  sealedIssueKeys: string[],
+): FakeResult[] {
+  const issueRows = issues as Array<{ repo: string; updated_at?: string | null }>;
+  if (sql.toLowerCase().includes("group by repo")) {
+    return aggregateRepoActivity(issueRows);
+  }
+  return dispatchByTable(sql, {
+    zk_opt_in: [{ zk_opt_in: zkOptIn ? 1 : 0 }],
+    "from zk_payloads": sealedIssueKeys.map((issue_key) => ({ issue_key })),
+    tenant_repo_access: visibleRepos.map((repo) => ({
+      repo,
+      is_private: 0,
+    })),
+    "from labels": labels as FakeResult[],
+    "from pr_state": prState as FakeResult[],
+    "from edges": edges as FakeResult[],
+    "from repos": repos as FakeResult[],
+    "from issues": issues as FakeResult[],
+  });
+}
+
+export function makeGraphEnv(
+  labels: unknown[],
+  prState: unknown[],
+  issues: unknown[],
+  edges: unknown[],
+  repos: unknown[] = [],
+  overrideVisible?: string[],
+  zkOptIn = false,
+  sealedIssueKeys: string[] = [],
+): Env {
+  const visibleRepos = overrideVisible ?? [
+    ...new Set((issues as Array<{ repo: string }>).map((i) => i.repo)),
+  ];
+  const db = makeFakeDb((sql, args) =>
+    makeFakeStmt(
+      sql,
+      args,
+      graphDbRows(
+        sql,
+        labels,
+        prState,
+        issues,
+        edges,
+        repos,
+        visibleRepos,
+        zkOptIn,
+        sealedIssueKeys,
+      ),
+      0,
+    ),
+  );
+  return makeEnv(db);
+}
+
+export function makeGraphEnvWithCapture(
+  labels: unknown[],
+  prState: unknown[],
+  issues: unknown[],
+  edges: unknown[],
+  repos: unknown[] = [],
+): { env: Env; capturedSqls: string[] } {
+  const capturedSqls: string[] = [];
+  const visibleRepos = [...new Set((issues as Array<{ repo: string }>).map((i) => i.repo))];
+  const { db } = captureDb((sql, _args) => {
+    capturedSqls.push(sql);
+    return graphDbRows(sql, labels, prState, issues, edges, repos, visibleRepos, false, []);
+  });
+  return { env: makeEnv(db), capturedSqls };
 }
