+++ worker/src/sync/sync.test.ts
@@ -1,5 +1,5 @@
+import "./sync-test-mocks";
 import { afterEach, describe, expect, it, vi } from "vitest";
-import type { Env } from "../types";
 import {
   BRANCH_ISSUE_RE,
   UPSERT_ISSUE_SQL,
@@ -17,11 +17,9 @@ import {
   incrementAuthFailures,
   isHalted,
   resetAuthFailures,
-  runSync,
   syncBranches,
   syncRepoBundle,
   syncRepoIssues,
-  writeRunAudit,
 } from "./sync";
 import type { EdgeData } from "./sync";
 
@@ -581,54 +579,6 @@ describe("extractFromLabels priority aliases", () => {
 // syncRepoIssues — basic coverage
 // ---------------------------------------------------------------------------
 
-vi.mock("./graphql", () => ({
-  ghGraphql: vi.fn(),
-  GraphQLError: class GraphQLError extends Error {
-    isAuth: boolean;
-    constructor(msg: string, isAuth = false) {
-      super(msg);
-      this.isAuth = isAuth;
-    }
-  },
-}));
-
-vi.mock("./queries", () => ({
-  ARCHIVED_REPOS_QUERY: "ARCHIVED_REPOS_QUERY",
-  ISSUES_QUERY: "ISSUES_QUERY",
-  ISSUES_QUERY_STRUCTURE_ONLY: "ISSUES_QUERY_STRUCTURE_ONLY",
-  PRS_QUERY: "PRS_QUERY",
-  REFS_QUERY: "REFS_QUERY",
-  REPO_BUNDLE_QUERY: "REPO_BUNDLE_QUERY",
-  REPO_BUNDLE_QUERY_STRUCTURE_ONLY: "REPO_BUNDLE_QUERY_STRUCTURE_ONLY",
-  REPOS_QUERY: "REPOS_QUERY",
-  STUB_ISSUE_QUERY: "STUB_ISSUE_QUERY",
-  STUB_ISSUE_QUERY_STRUCTURE_ONLY: "STUB_ISSUE_QUERY_STRUCTURE_ONLY",
-  pickIssuesQuery: (structureOnly: boolean) =>
-    structureOnly ? "ISSUES_QUERY_STRUCTURE_ONLY" : "ISSUES_QUERY",
-  pickRepoBundleQuery: (structureOnly: boolean) =>
-    structureOnly ? "REPO_BUNDLE_QUERY_STRUCTURE_ONLY" : "REPO_BUNDLE_QUERY",
-  pickStubIssueQuery: (structureOnly: boolean) =>
-    structureOnly ? "STUB_ISSUE_QUERY_STRUCTURE_ONLY" : "STUB_ISSUE_QUERY",
-}));
-
-vi.mock("../auth/installToken", () => ({
-  getInstallationToken: vi.fn(),
-  resolveInstallToken: vi.fn(),
-  listInstallationRepos: vi.fn(),
-}));
-
-vi.mock("./repo-probe", async (importOriginal) => {
-  const actual = await importOriginal<typeof import("./repo-probe")>();
-  return {
-    ...actual,
-    filterResolvableRepos: vi.fn(async (_token: string, repos: unknown[]) => ({
-      kept: repos,
-      dropped: [] as string[],
-    })),
-    isRepoResolvable: vi.fn().mockResolvedValue(true),
-  };
-});
-
 describe("syncRepoIssues", () => {
   it("upserts issues, deletes+inserts labels, collects edges, writes sync_state", async () => {
     const { ghGraphql } = await import("./graphql");
@@ -1075,1204 +1025,6 @@ describe("syncRepoBundle", () => {
     expect(closed).toBe(1);
   });
 });
-
-// ---------------------------------------------------------------------------
-// runSync — integration branches
-// ---------------------------------------------------------------------------
-
-describe("runSync", () => {
-  function makeEnv(overrides: Record<string, unknown> = {}): Env {
-    return {
-      DB: undefined as unknown as D1Database,
-      ...overrides,
-    } as unknown as Env;
-  }
-
-  it("returns early when isHalted=true", async () => {
-    const db = makeFakeDb((sql, args) => {
-      // All queries return halted=1
-      return makeFakeStmt(sql, args, [{ value: "1" }], 0);
-    });
-    const env = makeEnv({ DB: db });
-
-    // Should not throw; acquireSyncLock must not be called (we return before it)
-    await expect(runSync(env)).resolves.toBeUndefined();
-
-    // releaseSyncLock is in finally; acquireSyncLock UPDATE → changes=0 → not acquired
-    // The key signal: ghGraphql was never called (mocked, no calls expected)
-    const { ghGraphql } = await import("./graphql");
-    expect(vi.mocked(ghGraphql)).not.toHaveBeenCalled();
-  });
-
-  it("returns early when lock not acquired", async () => {
-    let callIdx = 0;
-    const db = makeFakeDb((sql, args) => {
-      callIdx++;
-      // First call: isHalted SELECT → halted=0
-      if (callIdx === 1) return makeFakeStmt(sql, args, [{ value: "0" }], 0);
-      // Second call: acquireSyncLock UPDATE → changes=0 (lock not acquired)
-      return makeFakeStmt(sql, args, [], 0);
-    });
-    const env = makeEnv({ DB: db });
-
-    await expect(runSync(env)).resolves.toBeUndefined();
-
-    const { ghGraphql } = await import("./graphql");
-    expect(vi.mocked(ghGraphql)).not.toHaveBeenCalled();
-  });
-
-  it("returns early (logs warn) when no tenants with installation_id exist", async () => {
-    const { listInstallationRepos } = await import("../auth/installToken");
-    const { getInstallationToken } = await import("../auth/installToken");
-    vi.mocked(getInstallationToken).mockResolvedValue("fake-token");
-    vi.mocked(listInstallationRepos).mockResolvedValue([]);
-
-    const db = makeFakeDb((sql, args) => {
-      if (sql.includes("FROM tenants")) {
-        // discoverTenants → empty (no tenants with installation_id)
-        return makeFakeStmt(sql, args, []);
-      }
-      return makeFakeStmt(sql, args, [], 1);
-    });
-
-    const env = makeEnv({ DB: db });
-
-    const warnSpy = vi.spyOn(console, "warn").mockImplementation(() => {});
-    await expect(runSync(env)).resolves.toBeUndefined();
-    expect(warnSpy).toHaveBeenCalledWith(
-      expect.stringContaining("no repos discovered across all installations — nothing to sync"),
-    );
-    warnSpy.mockRestore();
-  });
-
-  // Prune + archived-repo tracking (#N) -------------------------------------
-
-  /**
-   * Builds a FakeD1 that drives runSync through the prune path:
-   *   discoverTenants → [{id:1, installation_id:139542392}] →
-   *   lock acquired → getInstallationToken → listInstallationRepos=[Roxabi/roxabi-factory] →
-   *   DB SELECT DISTINCT queries return stale=[Roxabi/lyra] in issues/edges/pr_state/sync_state →
-   *   batchChunked invoked with DELETE stmts for Roxabi/lyra
-   */
-  function makeFullSyncDb(
-    opts: {
-      issueRepos?: string[];
-      edgeSrcRepos?: string[];
-      edgeDstRepos?: string[];
-      prStateRepos?: string[];
-      syncStateRepos?: string[];
-    } = {},
-  ) {
-    const {
-      issueRepos = ["Roxabi/lyra", "Roxabi/roxabi-factory"],
-      edgeSrcRepos = ["Roxabi/lyra"],
-      edgeDstRepos = [],
-      prStateRepos = ["Roxabi/lyra"],
-      syncStateRepos = ["Roxabi/lyra", "Roxabi/roxabi-factory"],
-    } = opts;
-
-    return makeFakeDb((sql, args) => {
-      if (sql.includes("key='halted'")) return makeFakeStmt(sql, args, [{ value: "0" }], 0);
-      if (sql.includes("sync_running") && sql.includes("UPDATE"))
-        return makeFakeStmt(sql, args, [], 1);
-      if (sql.includes("sync_started_at")) return makeFakeStmt(sql, args, [], 1);
-      // discoverTenants
-      if (sql.includes("FROM tenants")) {
-        return makeFakeStmt(sql, args, [{ id: 1, installation_id: 139542392 }]);
-      }
-      // tenant_repo_access (not used in prune path but respond gracefully)
-      if (sql.includes("FROM tenant_repo_access")) {
-        return makeFakeStmt(sql, args, [{ owner: "Roxabi", name: "roxabi-factory" }]);
-      }
-      // Prune SELECT DISTINCT queries
-      if (sql.includes("SELECT DISTINCT repo FROM issues")) {
-        return makeFakeStmt(
-          sql,
-          args,
-          issueRepos.map((r) => ({ repo: r })),
-        );
-      }
-      if (sql.includes("SELECT DISTINCT substr(src_key")) {
-        return makeFakeStmt(
-          sql,
-          args,
-          edgeSrcRepos.map((r) => ({ repo: r })),
-        );
-      }
-      if (sql.includes("SELECT DISTINCT substr(dst_key")) {
-        return makeFakeStmt(
-          sql,
-          args,
-          edgeDstRepos.map((r) => ({ repo: r })),
-        );
-      }
-      if (sql.includes("SELECT DISTINCT repo FROM pr_state")) {
-        return makeFakeStmt(
-          sql,
-          args,
-          prStateRepos.map((r) => ({ repo: r })),
-        );
-      }
-      if (sql.includes("SELECT repo FROM sync_state")) {
-        return makeFakeStmt(
-          sql,
-          args,
-          syncStateRepos.map((r) => ({ repo: r })),
-        );
-      }
-      // releaseSyncLock, writeRunAudit queries
-      return makeFakeStmt(sql, args, [], 1);
-    });
-  }
-
-  it("prunes issues/edges/pr_state/sync_state for a deleted repo (Roxabi/lyra)", async () => {
-    const { getInstallationToken, listInstallationRepos } = await import("../auth/installToken");
-    vi.mocked(getInstallationToken).mockResolvedValue("fake-token");
-    // listInstallationRepos returns only roxabi-factory (lyra is gone)
-    vi.mocked(listInstallationRepos).mockResolvedValue([
-      { repo: "Roxabi/roxabi-factory", isPrivate: false },
-    ]);
-
-    const { ghGraphql } = await import("./graphql");
-    const mockGhGraphql = vi.mocked(ghGraphql);
-    // syncRepoBundle calls for roxabi-factory — empty result
-    mockGhGraphql.mockResolvedValue({
-      data: {
-        repository: {
-          issues: { pageInfo: { hasNextPage: false, endCursor: null }, nodes: [] },
-          refs: { pageInfo: { hasNextPage: false, endCursor: null }, nodes: [] },
-          pullRequests: { pageInfo: { hasNextPage: false, endCursor: null }, nodes: [] },
-        },
-        rateLimit: { cost: 1, remaining: 4998, resetAt: "2026-01-01T00:00:00Z" },
-      },
-    });
-
-    const db = makeFullSyncDb();
-    vi.spyOn(console, "log").mockImplementation(() => {});
-
-    const env = makeEnv({ DB: db });
-    await expect(runSync(env)).resolves.toBeUndefined();
-
-    // Verify DELETE FROM issues was issued for Roxabi/lyra
-    const deletedIssueStmts = db._recorded.filter(
-      (s) => s.sql.includes("DELETE FROM issues") && s.args.includes("Roxabi/lyra"),
-    );
-    expect(deletedIssueStmts.length).toBeGreaterThan(0);
-
-    // Verify DELETE FROM edges was issued for Roxabi/lyra
-    const deletedEdgeStmts = db._recorded.filter(
-      (s) => s.sql.includes("DELETE FROM edges") && s.args.includes("Roxabi/lyra"),
-    );
-    expect(deletedEdgeStmts.length).toBeGreaterThan(0);
-  });
-
-  it("does NOT prune rows for a repo still accessible via installation (roxabi-vault stays)", async () => {
-    const { getInstallationToken, listInstallationRepos } = await import("../auth/installToken");
-    vi.mocked(getInstallationToken).mockResolvedValue("fake-token");
-    // listInstallationRepos returns both repos — roxabi-vault is still accessible
-    vi.mocked(listInstallationRepos).mockResolvedValue([
-      { repo: "Roxabi/roxabi-factory", isPrivate: false },
-      { repo: "Roxabi/roxabi-vault", isPrivate: false },
-    ]);
-
-    const { ghGraphql } = await import("./graphql");
-    const mockGhGraphql = vi.mocked(ghGraphql);
-    // syncRepoBundle empty result for each repo
-    mockGhGraphql.mockResolvedValue({
-      data: {
-        repository: {
-          issues: { pageInfo: { hasNextPage: false, endCursor: null }, nodes: [] },
-          refs: { pageInfo: { hasNextPage: false, endCursor: null }, nodes: [] },
-          pullRequests: { pageInfo: { hasNextPage: false, endCursor: null }, nodes: [] },
-        },
-        rateLimit: { cost: 1, remaining: 4998, resetAt: "2026-01-01T00:00:00Z" },
-      },
-    });
-
-    // DB returns roxabi-vault as present in issues — but it's still in live set, so NOT stale
-    const db = makeFullSyncDb({
-      issueRepos: ["Roxabi/roxabi-vault", "Roxabi/roxabi-factory"],
-      edgeSrcRepos: [],
-      edgeDstRepos: [],
-      prStateRepos: [],
-      syncStateRepos: ["Roxabi/roxabi-vault", "Roxabi/roxabi-factory"],
-    });
-    vi.spyOn(console, "log").mockImplementation(() => {});
-
-    const env = makeEnv({ DB: db });
-    await expect(runSync(env)).resolves.toBeUndefined();
-
-    // No DELETE statements should target roxabi-vault
-    const deletedVaultStmts = db._recorded.filter(
-      (s) => s.sql.includes("DELETE") && s.args.includes("Roxabi/roxabi-vault"),
-    );
-    expect(deletedVaultStmts).toHaveLength(0);
-  });
-
-  it("skips all prune logic (warn) when listInstallationRepos returns 0 repos", async () => {
-    const { getInstallationToken, listInstallationRepos } = await import("../auth/installToken");
-    vi.mocked(getInstallationToken).mockResolvedValue("fake-token");
-    // listInstallationRepos returns empty — transient error
-    vi.mocked(listInstallationRepos).mockResolvedValue([]);
-
-    const db = makeFakeDb((sql, args) => {
-      if (sql.includes("key='halted'")) return makeFakeStmt(sql, args, [{ value: "0" }], 0);
-      if (sql.includes("sync_running") && sql.includes("UPDATE"))
-        return makeFakeStmt(sql, args, [], 1);
-      if (sql.includes("sync_started_at")) return makeFakeStmt(sql, args, [], 1);
-      if (sql.includes("FROM tenants")) {
-        return makeFakeStmt(sql, args, [{ id: 1, installation_id: 139542392 }]);
-      }
-      return makeFakeStmt(sql, args, [], 1);
-    });
-
-    const warnSpy = vi.spyOn(console, "warn").mockImplementation(() => {});
-    vi.spyOn(console, "log").mockImplementation(() => {});
-
-    const env = makeEnv({ DB: db });
-    await expect(runSync(env)).resolves.toBeUndefined();
-
-    // Safety guard must have warned — new flow emits the "nothing to sync" message
-    expect(warnSpy).toHaveBeenCalledWith(
-      expect.stringContaining("no repos discovered across all installations — nothing to sync"),
-    );
-
-    // No DELETE statements should have been issued
-    const deleteStmts = db._recorded.filter((s) => s.sql.includes("DELETE"));
-    expect(deleteStmts).toHaveLength(0);
-  });
-});
-
-// ---------------------------------------------------------------------------
-// writeRunAudit (#120)
-// ---------------------------------------------------------------------------
-
-describe("writeRunAudit", () => {
-  function auditDb() {
-    return makeFakeDb((sql, args) => {
-      if (sql.includes("FROM issues")) return makeFakeStmt(sql, args, [{ c: 2650 }]);
-      if (sql.includes("FROM edges")) return makeFakeStmt(sql, args, [{ c: 2432 }]);
-      if (sql.includes("FROM pr_state")) return makeFakeStmt(sql, args, [{ c: 373 }]);
-      if (sql.includes("MAX(last_synced_at)"))
-        return makeFakeStmt(sql, args, [{ w: "2026-06-08T09:00:00Z" }]);
-      if (sql.includes("sync_control"))
-        return makeFakeStmt(sql, args, [
-          { key: "halted", value: "0" },
-          { key: "auth_failures", value: "0" },
-        ]);
-      return makeFakeStmt(sql, args, []);
-    });
-  }
-
-  it("no-ops when LOGS is unbound (never throws)", async () => {
-    const env = { DB: auditDb() } as unknown as Env;
-    await expect(
-      writeRunAudit(env, env.DB, { outcome: "success", stubs: 0, durationMs: 1 }),
-    ).resolves.toBeUndefined();
-  });
-
-  it("puts a JSON snapshot with counts + watermark when LOGS is bound", async () => {
-    vi.spyOn(console, "log").mockImplementation(() => {});
-    const put = vi.fn().mockResolvedValue(undefined);
-    const db = auditDb();
-    const env = { DB: db, LOGS: { put } } as unknown as Env;
-
-    await writeRunAudit(env, db, { outcome: "success", stubs: 4, durationMs: 1234 });
-
-    expect(put).toHaveBeenCalledTimes(1);
-    const [key, body, opts] = put.mock.calls[0];
-    expect(key).toMatch(/^runs\/\d{4}-\d{2}-\d{2}\/.+\.json$/);
-    expect(opts).toMatchObject({ httpMetadata: { contentType: "application/json" } });
-    const snap = JSON.parse(body as string);
-    expect(snap).toMatchObject({
-      outcome: "success",
-      stubs: 4,
-      durationMs: 1234,
-      issues: 2650,
-      edges: 2432,
-      prs: 373,
-      watermark: "2026-06-08T09:00:00Z",
-      halted: false,
-      authFailures: 0,
-    });
-  });
-
-  it("records net deltas + corrections when before-snapshot + corrections supplied (#80)", async () => {
-    vi.spyOn(console, "log").mockImplementation(() => {});
-    const put = vi.fn().mockResolvedValue(undefined);
-    const db = auditDb();
-    const env = { DB: db, LOGS: { put } } as unknown as Env;
-
-    await writeRunAudit(env, db, {
-      outcome: "success",
-      stubs: 2,
-      durationMs: 500,
-      before: { issues: 2600, edges: 2400, prs: 370 },
-      reposSynced: 8,
-      reposSkipped: 1,
-      corrections: { stalePrsClosed: 3, staleReposPruned: 1, staleTenantReposRemoved: 2 },
-    });
-
-    const snap = JSON.parse(put.mock.calls[0][1] as string);
-    // after (auditDb) = {2650, 2432, 373}; before = {2600, 2400, 370}
-    expect(snap.deltas).toEqual({ issues: 50, edges: 32, prs: 3 });
-    expect(snap.reposSynced).toBe(8);
-    expect(snap.reposSkipped).toBe(1);
-    expect(snap.corrections).toEqual({
-      stubsCreated: 2,
-      stalePrsClosed: 3,
-      staleReposPruned: 1,
-      staleTenantReposRemoved: 2,
-    });
-  });
-
-  it("emits zero deltas when no before-snapshot is supplied", async () => {
-    vi.spyOn(console, "log").mockImplementation(() => {});
-    const put = vi.fn().mockResolvedValue(undefined);
-    const db = auditDb();
-    const env = { DB: db, LOGS: { put } } as unknown as Env;
-
-    await writeRunAudit(env, db, { outcome: "success", stubs: 0, durationMs: 1 });
-
-    const snap = JSON.parse(put.mock.calls[0][1] as string);
-    expect(snap.deltas).toEqual({ issues: 0, edges: 0, prs: 0 });
-    expect(snap.corrections).toEqual({
-      stubsCreated: 0,
-      stalePrsClosed: 0,
-      staleReposPruned: 0,
-      staleTenantReposRemoved: 0,
-    });
-  });
-
-  it("swallows R2 put failures (audit must not fail the sync)", async () => {
-    vi.spyOn(console, "error").mockImplementation(() => {});
-    const put = vi.fn().mockRejectedValue(new Error("R2 down"));
-    const db = auditDb();
-    const env = { DB: db, LOGS: { put } } as unknown as Env;
-
-    await expect(
-      writeRunAudit(env, db, { outcome: "error", stubs: 0, durationMs: 5 }),
-    ).resolves.toBeUndefined();
-    expect(put).toHaveBeenCalledTimes(1);
-  });
-});
-
-// ---------------------------------------------------------------------------
-// runSync — multi-tenant / per-installation tests
-// ---------------------------------------------------------------------------
-// DEFERRED to #160 (S3b — sync per-tenant cutover + PAT retirement). The runSync
-// rewrite was split out of #146 (S3a shipped the install-token infra + webhook
-// cutover). Skipped here because (a) these assert the not-yet-built per-tenant
-// runSync, and (b) two bugs must be fixed when un-skipping in #160:
-//   1. dedup assertion uses c[2]/c[3]; ghGraphql(query, variables, token) puts
-//      query at c[0], vars at c[1].
-//   2. no vi.mock("../auth/installToken") → a faithful impl's getInstallationToken
-//      throws before syncRepoBundle. Add the mock when implementing #160.
-// ---------------------------------------------------------------------------
-
-describe("runSync — multi-tenant installation sync (#160)", () => {
-  // Local makeEnv scoped to this describe so it composes cleanly
-  function makeEnv(overrides: Record<string, unknown> = {}): Env {
-    return {
-      DB: undefined as unknown as D1Database,
-      ...overrides,
-    } as unknown as Env;
-  }
-
-  it("deduplicates GraphQL bundle fetches: two tenants sharing repo o/r → exactly 1 REPO_BUNDLE_QUERY issued for o/r", async () => {
-    // Two tenants both have access to "o/r" via their install tokens.
-    // runSync must deduplicate: one syncRepoBundle call regardless of tenant count.
-    const { getInstallationToken, listInstallationRepos } = await import("../auth/installToken");
-    vi.mocked(getInstallationToken).mockResolvedValue("fake-token");
-    // Both tenants enumerate the same repo "o/r"
-    vi.mocked(listInstallationRepos).mockResolvedValue([{ repo: "o/r", isPrivate: false }]);
-
-    const db = makeFakeDb((sql, args) => {
-      if (sql.includes("key='halted'")) return makeFakeStmt(sql, args, [{ value: "0" }], 0);
-      if (sql.includes("sync_running") && sql.includes("UPDATE"))
-        return makeFakeStmt(sql, args, [], 1);
-      if (sql.includes("sync_started_at")) return makeFakeStmt(sql, args, [], 1);
-      if (sql.includes("FROM tenants")) {
-        return makeFakeStmt(sql, args, [
-          { id: 1, installation_id: 10 },
-          { id: 2, installation_id: 20 },
-        ]);
-      }
-      return makeFakeStmt(sql, args, [], 1);
-    });
-
-    const { ghGraphql } = await import("./graphql");
-    const mockGhGraphql = vi.mocked(ghGraphql);
-    // syncRepoBundle uses REPO_BUNDLE_QUERY — stub returns minimal valid shape
-    mockGhGraphql.mockResolvedValue({
-      data: {
-        repository: {
-          issues: { pageInfo: { hasNextPage: false, endCursor: null }, nodes: [] },
-          refs: { pageInfo: { hasNextPage: false, endCursor: null }, nodes: [] },
-          pullRequests: { pageInfo: { hasNextPage: false, endCursor: null }, nodes: [] },
-        },
-        rateLimit: { cost: 1, remaining: 4999, resetAt: "2026-01-01T00:00:00Z" },
-      },
-    });
-
-    const env = makeEnv({ DB: db });
-    vi.spyOn(console, "log").mockImplementation(() => {});
-    await runSync(env);
-
-    // Exactly 1 REPO_BUNDLE_QUERY for o/r (probe calls during discovery are separate)
-    const bundleCalls = mockGhGraphql.mock.calls.filter((c) => c[0] === "REPO_BUNDLE_QUERY");
-    expect(bundleCalls).toHaveLength(1);
-  });
-
-  it("advances sync_slot per tick: seed sync_slot=0, after runSync the slot is updated", async () => {
-    // sync_control row sync_slot tracks which window of repos was processed.
-    // Discovery must yield ≥1 repo so Phase 2 runs and reaches the slot-advance write.
-    const { getInstallationToken, listInstallationRepos } = await import("../auth/installToken");
-    vi.mocked(getInstallationToken).mockResolvedValue("fake-token");
-    vi.mocked(listInstallationRepos).mockResolvedValue([{ repo: "o/r", isPrivate: false }]);
-
-    const { ghGraphql } = await import("./graphql");
-    vi.mocked(ghGraphql).mockResolvedValue({
-      data: {
-        repository: {
-          issues: { pageInfo: { hasNextPage: false, endCursor: null }, nodes: [] },
-          refs: { pageInfo: { hasNextPage: false, endCursor: null }, nodes: [] },
-          pullRequests: { pageInfo: { hasNextPage: false, endCursor: null }, nodes: [] },
-        },
-        rateLimit: { cost: 1, remaining: 4999, resetAt: "2026-01-01T00:00:00Z" },
-      },
-    });
-
-    let syncSlotWritten = false;
-    let nextSlotValue: unknown;
-    const db = makeFakeDb((sql, args) => {
-      if (sql.includes("key='halted'")) return makeFakeStmt(sql, args, [{ value: "0" }], 0);
-      if (sql.includes("sync_running") && sql.includes("UPDATE"))
-        return makeFakeStmt(sql, args, [], 1);
-      if (sql.includes("sync_started_at")) return makeFakeStmt(sql, args, [], 1);
-      if (sql.includes("FROM tenants")) {
-        return makeFakeStmt(sql, args, [{ id: 1, installation_id: 10 }]);
-      }
-      // Return sync_slot=0 on read
-      if (sql.includes("sync_slot") && sql.includes("SELECT")) {
-        return makeFakeStmt(sql, args, [{ value: "0" }]);
-      }
-      // Detect the slot-advance write (UPDATE with key=sync_slot); capture bound value.
-      if (sql.includes("sync_slot") && sql.includes("UPDATE")) {
-        syncSlotWritten = true;
-        nextSlotValue = args[0];
-        return makeFakeStmt(sql, args, [], 1);
-      }
-      return makeFakeStmt(sql, args, [], 1);
-    });
-
-    vi.spyOn(console, "log").mockImplementation(() => {});
-    vi.spyOn(console, "warn").mockImplementation(() => {});
-    const env = makeEnv({ DB: db });
-    await runSync(env);
-
-    // The sync run must advance sync_slot: 0 → (0+1) % NUM_SLOTS = 1.
-    expect(syncSlotWritten).toBe(true);
-    expect(nextSlotValue).toBe("1");
-  });
-
-  it("per-tenant lock isolation: tenant A holding sync_running=1 does NOT prevent tenant B from syncing", async () => {
-    // acquireSyncLock(db, tenantId) is per-tenant: tenant A's held lock (changes=0)
-    // must not stop discovery from attempting tenant B's own lock.
-    const { getInstallationToken, listInstallationRepos } = await import("../auth/installToken");
-    vi.mocked(getInstallationToken).mockResolvedValue("fake-token");
-    vi.mocked(listInstallationRepos).mockResolvedValue([]);
-
-    let tenantBLockAttempted = false;
-    const db = makeFakeDb((sql, args) => {
-      if (sql.includes("key='halted'")) return makeFakeStmt(sql, args, [{ value: "0" }], 0);
-      if (sql.includes("sync_running") && sql.includes("UPDATE")) {
-        // acquireSyncLock binds: .bind(timestamp, tenantId)
-        // → args[0] = ISO timestamp (string), args[1] = tenantId (number).
-        // CONTRACT: if the lock UPDATE bind order changes, args[1] becomes the ISO
-        // timestamp string → Number(...) is NaN → none of the tenantId branches match →
-        // tenantBLockAttempted stays false → the final assertion fails loudly.
-        const tenantId = Number(args[1]);
-        if (tenantId === 0) return makeFakeStmt(sql, args, [], 1); // global tick lock: acquired
-        if (tenantId === 1) return makeFakeStmt(sql, args, [], 0); // tenant A: lock held
-        if (tenantId === 2) {
-          tenantBLockAttempted = true;
-          return makeFakeStmt(sql, args, [], 1); // tenant B: lock acquired
-        }
-        return makeFakeStmt(sql, args, [], 0);
-      }
-      if (sql.includes("sync_started_at")) return makeFakeStmt(sql, args, [], 1);
-      if (sql.includes("FROM tenants")) {
-        return makeFakeStmt(sql, args, [
-          { id: 1, installation_id: 10 },
-          { id: 2, installation_id: 20 },
-        ]);
-      }
-      return makeFakeStmt(sql, args, [], 1);
-    });
-
-    const warnSpy = vi.spyOn(console, "warn").mockImplementation(() => {});
-    const env = makeEnv({ DB: db });
-    await runSync(env);
-    warnSpy.mockRestore();
-
-    // Tenant B's lock attempt must have been made (runSync iterates tenants independently).
-    expect(tenantBLockAttempted).toBe(true);
-  });
-
-  it("no PAT access: runSync completes via install tokens without reading env.GITHUB_TOKEN", async () => {
-    // runSync must use getInstallationToken() and never read env.GITHUB_TOKEN.
-    const { getInstallationToken, listInstallationRepos } = await import("../auth/installToken");
-    vi.mocked(getInstallationToken).mockResolvedValue("fake-token");
-    vi.mocked(listInstallationRepos).mockResolvedValue([
-      { repo: "Roxabi/roxabi-factory", isPrivate: false },
-    ]);
-
-    const { ghGraphql } = await import("./graphql");
-    vi.mocked(ghGraphql).mockResolvedValue({
-      data: {
-        repository: {
-          issues: { pageInfo: { hasNextPage: false, endCursor: null }, nodes: [] },
-          refs: { pageInfo: { hasNextPage: false, endCursor: null }, nodes: [] },
-          pullRequests: { pageInfo: { hasNextPage: false, endCursor: null }, nodes: [] },
-        },
-        rateLimit: { cost: 1, remaining: 4999, resetAt: "2026-01-01T00:00:00Z" },
-      },
-    });
-
-    const db = makeFakeDb((sql, args) => {
-      if (sql.includes("key='halted'")) return makeFakeStmt(sql, args, [{ value: "0" }], 0);
-      if (sql.includes("sync_running") && sql.includes("UPDATE"))
-        return makeFakeStmt(sql, args, [], 1);
-      if (sql.includes("sync_started_at")) return makeFakeStmt(sql, args, [], 1);
-      if (sql.includes("FROM tenants")) {
-        return makeFakeStmt(sql, args, [{ id: 1, installation_id: 139542392 }]);
-      }
-      return makeFakeStmt(sql, args, [], 1);
-    });
-
-    const base = makeEnv({ DB: db });
-    // Replace GITHUB_TOKEN with a spy getter that records every access.
-    let patAccessCount = 0;
-    (base as unknown as Record<string, unknown>).GITHUB_TOKEN = undefined;
-    Object.defineProperty(base, "GITHUB_TOKEN", {
-      get() {
-        patAccessCount++;
-        return "tok";
-      },
-      configurable: true,
-    });
-
-    vi.spyOn(console, "log").mockImplementation(() => {});
-    await runSync(base);
-
-    // runSync must not read env.GITHUB_TOKEN — install token is used exclusively.
-    expect(patAccessCount).toBe(0);
-  });
-
-  it("window-past-end: sync_slot beyond repo count → no REPO_BUNDLE_QUERY fetches and runSync resolves", async () => {
-    // WINDOW=20 (sync.ts const), NUM_SLOTS=2 → code only persists slots {0,1}.
-    // slot=2 here is a FORCED out-of-range value to exercise the slice-past-end
-    // guard defensively (a stale/garbage persisted slot must not crash Phase 2).
-    // 21 repos → windowing engages (allRepos.length > WINDOW).
-    // slot=2 → windowStart = 2 * 20 = 40 ≥ 21 → windowedRepos=[] → no bundle fetches.
-    const { getInstallationToken, listInstallationRepos } = await import("../auth/installToken");
-    vi.mocked(getInstallationToken).mockResolvedValue("fake-token");
-    // 21 unique repos to exceed WINDOW=20
-    vi.mocked(listInstallationRepos).mockResolvedValue(
-      Array.from({ length: 21 }, (_, i) => ({ repo: `o/r${i}`, isPrivate: false })),
-    );
-
-    const { ghGraphql } = await import("./graphql");
-    const mockGhGraphql = vi.mocked(ghGraphql);
-    mockGhGraphql.mockResolvedValue({
-      data: {
-        repository: {
-          issues: { pageInfo: { hasNextPage: false, endCursor: null }, nodes: [] },
-          refs: { pageInfo: { hasNextPage: false, endCursor: null }, nodes: [] },
-          pullRequests: { pageInfo: { hasNextPage: false, endCursor: null }, nodes: [] },
-        },
-        rateLimit: { cost: 1, remaining: 4999, resetAt: "2026-01-01T00:00:00Z" },
-      },
-    });
-
-    const db = makeFakeDb((sql, args) => {
-      if (sql.includes("key='halted'")) return makeFakeStmt(sql, args, [{ value: "0" }], 0);
-      if (sql.includes("sync_running") && sql.includes("UPDATE"))
-        return makeFakeStmt(sql, args, [], 1);
-      if (sql.includes("sync_started_at")) return makeFakeStmt(sql, args, [], 1);
-      if (sql.includes("FROM tenants")) {
-        return makeFakeStmt(sql, args, [{ id: 1, installation_id: 10 }]);
-      }
-      // Return slot=2: windowStart = 2*20 = 40 ≥ 21 repos → windowedRepos=[]
-      if (sql.includes("sync_slot") && sql.includes("SELECT")) {
-        return makeFakeStmt(sql, args, [{ value: "2" }]);
-      }
-      return makeFakeStmt(sql, args, [], 1);
-    });
-
-    vi.spyOn(console, "log").mockImplementation(() => {});
-    vi.spyOn(console, "warn").mockImplementation(() => {});
-    const env = makeEnv({ DB: db });
-
-    // Must resolve without throwing — windowed-past-end is a valid no-op for Phase 2.
-    await expect(runSync(env)).resolves.toBeUndefined();
-
-    // No REPO_BUNDLE_QUERY calls: Phase 2 was skipped entirely.
-    const bundleCalls = mockGhGraphql.mock.calls.filter((c) => c[0] === "REPO_BUNDLE_QUERY");
-    expect(bundleCalls).toHaveLength(0);
-  });
-});
-
-// ---------------------------------------------------------------------------
-// runSync — breaker + discovery (#160)
-// ---------------------------------------------------------------------------
-
-describe("runSync — breaker + discovery (#160)", () => {
-  function makeEnv(overrides: Record<string, unknown> = {}): Env {
-    return {
-      DB: undefined as unknown as D1Database,
-      ...overrides,
-    } as unknown as Env;
-  }
-
-  afterEach(() => {
-    vi.restoreAllMocks();
-  });
-
-  it("B1: discovery failure increments the tenant breaker", async () => {
-    // Arrange
-    const { getInstallationToken, listInstallationRepos } = await import("../auth/installToken");
-    vi.mocked(getInstallationToken).mockRejectedValue(new Error("token-error"));
-    vi.mocked(listInstallationRepos).mockResolvedValue([]);
-
-    const db = makeFakeDb((sql, args) => {
-      // Global tick lock (tenant_id=0) → acquired (changes=1)
-      if (sql.includes("sync_running") && sql.includes("UPDATE") && args[0] === 0) {
-        return makeFakeStmt(sql, args, [], 1);
-      }
-      // Tenant lock (tenant_id=1) → acquired
-      if (sql.includes("sync_running") && sql.includes("UPDATE")) {
-        return makeFakeStmt(sql, args, [], 1);
-      }
-      // halted guard
-      if (sql.includes("key='halted'") && sql.includes("SELECT")) {
-        return makeFakeStmt(sql, args, [{ value: "0" }], 0);
-      }
-      // auth_failures SELECT for tenant after increment → return count
-      if (
-        sql.includes("key='auth_failures'") &&
-        sql.includes("SELECT") &&
-        !sql.includes("UPDATE")
-      ) {
-        return makeFakeStmt(sql, args, [{ value: "1" }], 0);
-      }
-      // tenants discovery
-      if (sql.includes("FROM tenants")) {
-        return makeFakeStmt(sql, args, [{ id: 1, installation_id: 10 }]);
-      }
-      return makeFakeStmt(sql, args, [], 1);
-    });
-
-    vi.spyOn(console, "log").mockImplementation(() => {});
-    vi.spyOn(console, "error").mockImplementation(() => {});
-
-    const env = makeEnv({ DB: db });
-
-    // Act
-    await runSync(env);
-
-    // Assert — auth_failures UPDATE was issued bound to tenant_id=1
-    const recorded = db._recorded;
-    const authFailureUpdate = recorded.find(
-      (s) => s.sql.includes("auth_failures") && s.sql.includes("UPDATE") && s.args.includes(1),
-    );
-    expect(authFailureUpdate).toBeDefined();
-  });
-
-  it("B3: discoverTenants upserts repos and deletes stale tenant_repo_access", async () => {
-    // Arrange
-    const { getInstallationToken, listInstallationRepos } = await import("../auth/installToken");
-    vi.mocked(getInstallationToken).mockResolvedValue("tok");
-    vi.mocked(listInstallationRepos).mockResolvedValue([{ repo: "o/keep", isPrivate: true }]);
-
-    const { ghGraphql } = await import("./graphql");
-    vi.mocked(ghGraphql).mockResolvedValue({
-      data: {
-        repository: {
-          issues: { pageInfo: { hasNextPage: false, endCursor: null }, nodes: [] },
-          refs: { pageInfo: { hasNextPage: false, endCursor: null }, nodes: [] },
-          pullRequests: { pageInfo: { hasNextPage: false, endCursor: null }, nodes: [] },
-        },
-        rateLimit: { cost: 1, remaining: 4999, resetAt: "2026-01-01T00:00:00Z" },
-      },
-    });
-
-    const db = makeFakeDb((sql, args) => {
-      // Global tick lock → acquired
-      if (sql.includes("sync_running") && sql.includes("UPDATE") && args[0] === 0) {
-        return makeFakeStmt(sql, args, [], 1);
-      }
-      // Tenant lock → acquired
-      if (sql.includes("sync_running") && sql.includes("UPDATE")) {
-        return makeFakeStmt(sql, args, [], 1);
-      }
-      // halted guard
-      if (sql.includes("key='halted'") && sql.includes("SELECT")) {
-        return makeFakeStmt(sql, args, [{ value: "0" }], 0);
-      }
-      // auth_failures SELECT → not halted
-      if (
-        sql.includes("key='auth_failures'") &&
-        sql.includes("SELECT") &&
-        !sql.includes("UPDATE")
-      ) {
-        return makeFakeStmt(sql, args, [{ value: "0" }], 0);
-      }
-      // tenants discovery → 1 tenant
-      if (sql.includes("FROM tenants")) {
-        return makeFakeStmt(sql, args, [{ id: 1, installation_id: 10 }]);
-      }
-      // existing tenant_repo_access rows → keep + stale
-      if (sql.includes("SELECT repo FROM tenant_repo_access")) {
-        return makeFakeStmt(sql, args, [{ repo: "o/keep" }, { repo: "o/stale" }]);
-      }
-      return makeFakeStmt(sql, args, [], 1);
-    });
-
-    vi.spyOn(console, "log").mockImplementation(() => {});
-    vi.spyOn(console, "error").mockImplementation(() => {});
-
-    const env = makeEnv({ DB: db });
-
-    // Act
-    await runSync(env);
-
-    // Assert — INSERT for "o/keep"
-    const recorded = db._recorded;
-    const insertKeep = recorded.find(
-      (s) => s.sql.includes("INSERT INTO tenant_repo_access") && s.args.includes("o/keep"),
-    );
-    expect(insertKeep).toBeDefined();
-    // #148: is_private written (o/keep mocked private → bound 1; args = [tenantId, repo, is_private])
-    expect(insertKeep?.args[2]).toBe(1);
-
-    // Assert — DELETE for "o/stale" (not for "o/keep")
-    const deleteStale = recorded.find(
-      (s) => s.sql.includes("DELETE FROM tenant_repo_access") && s.args.includes("o/stale"),
-    );
-    expect(deleteStale).toBeDefined();
-
-    const deleteKeep = recorded.find(
-      (s) => s.sql.includes("DELETE FROM tenant_repo_access") && s.args.includes("o/keep"),
-    );
-    expect(deleteKeep).toBeUndefined();
-  });
-
-  it("B3b: discoverTenants writes is_private=0 for a public repo", async () => {
-    // Arrange
-    const { getInstallationToken, listInstallationRepos } = await import("../auth/installToken");
-    vi.mocked(getInstallationToken).mockResolvedValue("tok");
-    vi.mocked(listInstallationRepos).mockResolvedValue([{ repo: "o/pub", isPrivate: false }]);
-
-    const { ghGraphql } = await import("./graphql");
-    vi.mocked(ghGraphql).mockResolvedValue({
-      data: {
-        repository: {
-          issues: { pageInfo: { hasNextPage: false, endCursor: null }, nodes: [] },
-          refs: { pageInfo: { hasNextPage: false, endCursor: null }, nodes: [] },
-          pullRequests: { pageInfo: { hasNextPage: false, endCursor: null }, nodes: [] },
-        },
-        rateLimit: { cost: 1, remaining: 4999, resetAt: "2026-01-01T00:00:00Z" },
-      },
-    });
-
-    const db = makeFakeDb((sql, args) => {
-      // Global tick lock → acquired
-      if (sql.includes("sync_running") && sql.includes("UPDATE") && args[0] === 0) {
-        return makeFakeStmt(sql, args, [], 1);
-      }
-      // Tenant lock → acquired
-      if (sql.includes("sync_running") && sql.includes("UPDATE")) {
-        return makeFakeStmt(sql, args, [], 1);
-      }
-      // halted guard
-      if (sql.includes("key='halted'") && sql.includes("SELECT")) {
-        return makeFakeStmt(sql, args, [{ value: "0" }], 0);
-      }
-      // auth_failures SELECT → not halted
-      if (
-        sql.includes("key='auth_failures'") &&
-        sql.includes("SELECT") &&
-        !sql.includes("UPDATE")
-      ) {
-        return makeFakeStmt(sql, args, [{ value: "0" }], 0);
-      }
-      // tenants discovery → 1 tenant
-      if (sql.includes("FROM tenants")) {
-        return makeFakeStmt(sql, args, [{ id: 1, installation_id: 10 }]);
-      }
-      // existing tenant_repo_access rows → same as current (no stale)
-      if (sql.includes("SELECT repo FROM tenant_repo_access")) {
-        return makeFakeStmt(sql, args, [{ repo: "o/pub" }]);
-      }
-      return makeFakeStmt(sql, args, [], 1);
-    });
-
-    vi.spyOn(console, "log").mockImplementation(() => {});
-    vi.spyOn(console, "error").mockImplementation(() => {});
-
-    const env = makeEnv({ DB: db });
-
-    // Act
-    await runSync(env);
-
-    // Assert — INSERT for "o/pub" with is_private=0
-    const recorded = db._recorded;
-    const insertPub = recorded.find(
-      (s) => s.sql.includes("INSERT INTO tenant_repo_access") && s.args.includes("o/pub"),
-    );
-    expect(insertPub).toBeDefined();
-    // #148: is_private=0 for public repo (args = [tenantId, repo, is_private])
-    expect(insertPub?.args[2]).toBe(0);
-  });
-
-  it("B3c: repos table gets archived=1 for an installation-archived repo, 0 for a live one (#160 fallout)", async () => {
-    // Arrange — installation reports one live + one archived repo.
-    const { getInstallationToken, listInstallationRepos } = await import("../auth/installToken");
-    vi.mocked(getInstallationToken).mockResolvedValue("tok");
-    vi.mocked(listInstallationRepos).mockResolvedValue([
-      { repo: "o/live", isPrivate: false, isArchived: false },
-      { repo: "o/old", isPrivate: false, isArchived: true },
-    ]);
-
-    const { ghGraphql } = await import("./graphql");
-    vi.mocked(ghGraphql).mockResolvedValue({
-      data: {
-        repository: {
-          issues: { pageInfo: { hasNextPage: false, endCursor: null }, nodes: [] },
-          refs: { pageInfo: { hasNextPage: false, endCursor: null }, nodes: [] },
-          pullRequests: { pageInfo: { hasNextPage: false, endCursor: null }, nodes: [] },
-        },
-        rateLimit: { cost: 1, remaining: 4999, resetAt: "2026-01-01T00:00:00Z" },
-      },
-    });
-
-    const db = makeFakeDb((sql, args) => {
-      if (sql.includes("sync_running") && sql.includes("UPDATE") && args[0] === 0) {
-        return makeFakeStmt(sql, args, [], 1);
-      }
-      if (sql.includes("sync_running") && sql.includes("UPDATE")) {
-        return makeFakeStmt(sql, args, [], 1);
-      }
-      if (sql.includes("key='halted'") && sql.includes("SELECT")) {
-        return makeFakeStmt(sql, args, [{ value: "0" }], 0);
-      }
-      if (
-        sql.includes("key='auth_failures'") &&
-        sql.includes("SELECT") &&
-        !sql.includes("UPDATE")
-      ) {
-        return makeFakeStmt(sql, args, [{ value: "0" }], 0);
-      }
-      if (sql.includes("FROM tenants")) {
-        return makeFakeStmt(sql, args, [{ id: 1, installation_id: 10 }]);
-      }
-      if (sql.includes("SELECT repo FROM tenant_repo_access")) {
-        return makeFakeStmt(sql, args, [{ repo: "o/live" }, { repo: "o/old" }]);
-      }
-      return makeFakeStmt(sql, args, [], 1);
-    });
-
-    vi.spyOn(console, "log").mockImplementation(() => {});
-    vi.spyOn(console, "error").mockImplementation(() => {});
-
-    const env = makeEnv({ DB: db });
-
-    // Act
-    await runSync(env);
-
-    // Assert — repos upsert binds archived per repo (args = [repo, archived]).
-    const recorded = db._recorded;
-    const insertOld = recorded.find(
-      (s) => s.sql.includes("INSERT INTO repos") && s.args[0] === "o/old",
-    );
-    const insertLive = recorded.find(
-      (s) => s.sql.includes("INSERT INTO repos") && s.args[0] === "o/live",
-    );
-    expect(insertOld?.args[1]).toBe(1);
-    expect(insertLive?.args[1]).toBe(0);
-  });
-
-  it("B4: token-exhausted repo is skipped, runSync resolves without throwing", async () => {
-    // Arrange — getInstallationToken always rejects for Phase 2 lookups
-    const { getInstallationToken, listInstallationRepos } = await import("../auth/installToken");
-    vi.mocked(getInstallationToken).mockRejectedValue(new Error("no-token"));
-    vi.mocked(listInstallationRepos).mockResolvedValue([{ repo: "o/a", isPrivate: false }]);
-
-    const { ghGraphql } = await import("./graphql");
-    vi.mocked(ghGraphql).mockResolvedValue({
-      data: {
-        repository: {
-          issues: { pageInfo: { hasNextPage: false, endCursor: null }, nodes: [] },
-          refs: { pageInfo: { hasNextPage: false, endCursor: null }, nodes: [] },
-          pullRequests: { pageInfo: { hasNextPage: false, endCursor: null }, nodes: [] },
-        },
-        rateLimit: { cost: 1, remaining: 4999, resetAt: "2026-01-01T00:00:00Z" },
-      },
-    });
-
-    const db = makeFakeDb((sql, args) => {
-      if (sql.includes("sync_running") && sql.includes("UPDATE") && args[0] === 0) {
-        return makeFakeStmt(sql, args, [], 1);
-      }
-      if (sql.includes("sync_running") && sql.includes("UPDATE")) {
-        return makeFakeStmt(sql, args, [], 1);
-      }
-      if (sql.includes("key='halted'") && sql.includes("SELECT")) {
-        return makeFakeStmt(sql, args, [{ value: "0" }], 0);
-      }
-      if (
-        sql.includes("key='auth_failures'") &&
-        sql.includes("SELECT") &&
-        !sql.includes("UPDATE")
-      ) {
-        return makeFakeStmt(sql, args, [{ value: "1" }], 0);
-      }
-      if (sql.includes("FROM tenants")) {
-        return makeFakeStmt(sql, args, [{ id: 1, installation_id: 10 }]);
-      }
-      if (sql.includes("SELECT repo FROM tenant_repo_access")) {
-        return makeFakeStmt(sql, args, [{ repo: "o/a" }]);
-      }
-      return makeFakeStmt(sql, args, [], 1);
-    });
-
-    vi.spyOn(console, "log").mockImplementation(() => {});
-    vi.spyOn(console, "error").mockImplementation(() => {});
-
-    const env = makeEnv({ DB: db });
-
-    // Act + Assert — runSync resolves (does not throw)
-    await expect(runSync(env)).resolves.toBeUndefined();
-
-    // ghGraphql must NOT have been called with REPO_BUNDLE_QUERY for o/a
-    const bundleCalls = vi.mocked(ghGraphql).mock.calls.filter((c) => c[0] === "REPO_BUNDLE_QUERY");
-    expect(bundleCalls).toHaveLength(0);
-  });
-
-  it("B2: systemic failure halts and POSTs NOTIFY at threshold (≥2)", async () => {
-    // Arrange — discovery succeeds (listInstallationRepos returns repos), Phase 2
-    // token fetch always fails → every windowed repo skipped → systemic failure.
-    const { getInstallationToken, listInstallationRepos } = await import("../auth/installToken");
-    // Discovery Phase 1 calls: getInstallationToken (once, for the tenant) then listInstallationRepos.
-    // We need Phase 1 to succeed so repos enter the window, then Phase 2 token fetch to fail.
-    // Use mockResolvedValueOnce for the Phase 1 call, then reject for all subsequent.
-    vi.mocked(getInstallationToken)
-      .mockResolvedValueOnce("tok-discovery") // Phase 1 success
-      .mockRejectedValue(new Error("systemic-error")); // Phase 2 failures
-    vi.mocked(listInstallationRepos).mockResolvedValue([{ repo: "o/r", isPrivate: false }]);
-
-    const { ghGraphql } = await import("./graphql");
-    vi.mocked(ghGraphql).mockResolvedValue({
-      data: { repository: { id: "1" } },
-    });
-
-    const db = makeFakeDb((sql, args) => {
-      if (sql.includes("sync_running") && sql.includes("UPDATE") && args[0] === 0) {
-        return makeFakeStmt(sql, args, [], 1);
-      }
-      if (sql.includes("sync_running") && sql.includes("UPDATE")) {
-        return makeFakeStmt(sql, args, [], 1);
-      }
-      if (sql.includes("key='halted'") && sql.includes("SELECT")) {
-        return makeFakeStmt(sql, args, [{ value: "0" }], 0);
-      }
-      // auth_failures SELECT — return ≥2 for global (tenant_id=0) to trigger halt
-      if (
-        sql.includes("key='auth_failures'") &&
-        sql.includes("SELECT") &&
-        !sql.includes("UPDATE")
-      ) {
-        const val = args[0] === 0 ? "2" : "1";
-        return makeFakeStmt(sql, args, [{ value: val }], 0);
-      }
-      if (sql.includes("FROM tenants")) {
-        return makeFakeStmt(sql, args, [{ id: 1, installation_id: 10 }]);
-      }
-      if (sql.includes("SELECT repo FROM tenant_repo_access")) {
-        return makeFakeStmt(sql, args, [{ repo: "o/r" }]);
-      }
-      return makeFakeStmt(sql, args, [], 1);
-    });
-
-    const fetchMock = vi.fn().mockResolvedValue({ ok: true });
-    vi.stubGlobal("fetch", fetchMock);
-
-    vi.spyOn(console, "log").mockImplementation(() => {});
-    vi.spyOn(console, "error").mockImplementation(() => {});
-
-    const env = makeEnv({ DB: db, NOTIFY_URL: "https://notify.example.com/hook" });
-
-    // Act
-    await runSync(env);
-
-    // Assert — a halted-related UPDATE was fired (haltSync writes to sync_control)
-    // The SQL sets halted='1' or value='1' where key='halted'
-    const recorded = db._recorded;
-    const haltUpdate = recorded.find((s) => s.sql.includes("halted") && s.sql.includes("UPDATE"));
-    expect(haltUpdate).toBeDefined();
-
-    // Assert — fetch was called with NOTIFY_URL and body containing 'sync_halted'
-    const notifyCalls = (fetchMock as ReturnType<typeof vi.fn>).mock.calls.filter(
-      (c: unknown[]) => typeof c[0] === "string" && (c[0] as string).includes("notify.example.com"),
-    );
-    expect(notifyCalls.length).toBeGreaterThan(0);
-    const bodyArg = notifyCalls[0][1] as { body?: string };
-    expect(bodyArg?.body).toMatch(/sync_halted/);
-  });
-
-  it("B2: systemic failure below threshold does NOT halt (auth_failures=1)", async () => {
-    // Arrange — same as B2 threshold but auth_failures stays at 1 (below halt trigger)
-    const { getInstallationToken, listInstallationRepos } = await import("../auth/installToken");
-    vi.mocked(getInstallationToken)
-      .mockResolvedValueOnce("tok-discovery") // Phase 1 success
-      .mockRejectedValue(new Error("systemic-error")); // Phase 2 failures
-    vi.mocked(listInstallationRepos).mockResolvedValue([{ repo: "o/r", isPrivate: false }]);
-
-    const { ghGraphql } = await import("./graphql");
-    vi.mocked(ghGraphql).mockResolvedValue({
-      data: { repository: { id: "1" } },
-    });
-
-    const db = makeFakeDb((sql, args) => {
-      if (sql.includes("sync_running") && sql.includes("UPDATE") && args[0] === 0) {
-        return makeFakeStmt(sql, args, [], 1);
-      }
-      if (sql.includes("sync_running") && sql.includes("UPDATE")) {
-        return makeFakeStmt(sql, args, [], 1);
-      }
-      if (sql.includes("key='halted'") && sql.includes("SELECT")) {
-        return makeFakeStmt(sql, args, [{ value: "0" }], 0);
-      }
-      // auth_failures SELECT always returns 1 (below threshold — no halt)
-      if (
-        sql.includes("key='auth_failures'") &&
-        sql.includes("SELECT") &&
-        !sql.includes("UPDATE")
-      ) {
-        return makeFakeStmt(sql, args, [{ value: "1" }], 0);
-      }
-      if (sql.includes("FROM tenants")) {
-        return makeFakeStmt(sql, args, [{ id: 1, installation_id: 10 }]);
-      }
-      if (sql.includes("SELECT repo FROM tenant_repo_access")) {
-        return makeFakeStmt(sql, args, [{ repo: "o/r" }]);
-      }
-      return makeFakeStmt(sql, args, [], 1);
-    });
-
-    const fetchMock = vi.fn().mockResolvedValue({ ok: true });
-    vi.stubGlobal("fetch", fetchMock);
-
-    vi.spyOn(console, "log").mockImplementation(() => {});
-    vi.spyOn(console, "error").mockImplementation(() => {});
-
-    const env = makeEnv({ DB: db, NOTIFY_URL: "https://notify.example.com/hook" });
-
-    // Act
-    await runSync(env);
-
-    // Assert — NO halted UPDATE issued
-    const recorded = db._recorded;
-    const haltUpdate = recorded.find((s) => s.sql.includes("halted") && s.sql.includes("UPDATE"));
-    expect(haltUpdate).toBeUndefined();
-
-    // Assert — fetch was NOT called with NOTIFY_URL
-    const notifyCalls = (fetchMock as ReturnType<typeof vi.fn>).mock.calls.filter(
-      (c: unknown[]) => typeof c[0] === "string" && (c[0] as string).includes("notify.example.com"),
-    );
-    expect(notifyCalls).toHaveLength(0);
-  });
-
-  it("B1b: successful run resets auth_failures for both global (tenant_id=0) and participating tenant", async () => {
-    // Arrange — 1 tenant id=7, 1 repo synced OK
-    const { getInstallationToken, listInstallationRepos } = await import("../auth/installToken");
-    vi.mocked(getInstallationToken).mockResolvedValue("tok-7");
-    vi.mocked(listInstallationRepos).mockResolvedValue([{ repo: "o/ok", isPrivate: false }]);
-
-    const { ghGraphql } = await import("./graphql");
-    vi.mocked(ghGraphql).mockResolvedValue({
-      data: {
-        repository: {
-          issues: { pageInfo: { hasNextPage: false, endCursor: null }, nodes: [] },
-          refs: { pageInfo: { hasNextPage: false, endCursor: null }, nodes: [] },
-          pullRequests: { pageInfo: { hasNextPage: false, endCursor: null }, nodes: [] },
-        },
-        rateLimit: { cost: 1, remaining: 4999, resetAt: "2026-01-01T00:00:00Z" },
-      },
-    });
-
-    const db = makeFakeDb((sql, args) => {
-      if (sql.includes("sync_running") && sql.includes("UPDATE") && args[0] === 0) {
-        return makeFakeStmt(sql, args, [], 1);
-      }
-      if (sql.includes("sync_running") && sql.includes("UPDATE")) {
-        return makeFakeStmt(sql, args, [], 1);
-      }
-      if (sql.includes("key='halted'") && sql.includes("SELECT")) {
-        return makeFakeStmt(sql, args, [{ value: "0" }], 0);
-      }
-      if (
-        sql.includes("key='auth_failures'") &&
-        sql.includes("SELECT") &&
-        !sql.includes("UPDATE")
-      ) {
-        return makeFakeStmt(sql, args, [{ value: "0" }], 0);
-      }
-      if (sql.includes("FROM tenants")) {
-        return makeFakeStmt(sql, args, [{ id: 7, installation_id: 70 }]);
-      }
-      if (sql.includes("SELECT repo FROM tenant_repo_access")) {
-        return makeFakeStmt(sql, args, [{ repo: "o/ok" }]);
-      }
-      return makeFakeStmt(sql, args, [], 1);
-    });
-
-    vi.spyOn(console, "log").mockImplementation(() => {});
-    vi.spyOn(console, "error").mockImplementation(() => {});
-
-    const env = makeEnv({ DB: db });
-
-    // Act
-    await runSync(env);
-
-    // Assert — auth_failures reset UPDATE (value='0') for tenant_id=0
-    const recorded = db._recorded;
-    const resetGlobal = recorded.find(
-      (s) =>
-        s.sql.includes("auth_failures") &&
-        s.sql.includes("UPDATE") &&
-        s.sql.includes("value='0'") &&
-        s.args.includes(0),
-    );
-    expect(resetGlobal).toBeDefined();
-
-    // Assert — auth_failures reset UPDATE (value='0') for tenant_id=7
-    const resetTenant = recorded.find(
-      (s) =>
-        s.sql.includes("auth_failures") &&
-        s.sql.includes("UPDATE") &&
-        s.sql.includes("value='0'") &&
-        s.args.includes(7),
-    );
-    expect(resetTenant).toBeDefined();
-  });
-});
-
-// ---------------------------------------------------------------------------
-// Teardown
-// ---------------------------------------------------------------------------
 
 afterEach(() => {
   vi.restoreAllMocks();
